Fixed a bug where unicode specials in String types were causing JSON decoding to fail Added a contract argument to TrasferChain to allow custom tokens to be transferred

+++ eosswift/eos-http-rpc/model/history/request/GetActions.swift
@@ -2,10 +2,10 @@ import Foundation
 
 public struct GetActions : Encodable {
     public let account_name: String
-    public let pos: Int?
-    public let offset: Int?
+    public let pos: Int64?
+    public let offset: Int64?
 
-    public init(account_name: String, pos: Int?, offset: Int?) {
+    public init(account_name: String, pos: Int64?, offset: Int64?) {
         self.account_name = account_name
         self.pos = pos
         self.offset = offset
