+++ eosswift/eos-http-rpc/model/transaction/response/TransactionReceipt.swift
@@ -3,12 +3,12 @@ import Foundation
 public struct TransactionReceipt : Decodable {
     public let receiver: String
     public let act_digest: String
-    public let global_sequence: Int
-    public let recv_sequence: Int
-    public let code_sequence: Int
-    public let abi_sequence: Int
+    public let global_sequence: Int64
+    public let recv_sequence: Int64
+    public let code_sequence: Int64
+    public let abi_sequence: Int64
 
-    public init(receiver: String, act_digest: String, global_sequence: Int, recv_sequence: Int, code_sequence: Int, abi_sequence: Int) {
+    public init(receiver: String, act_digest: String, global_sequence: Int64, recv_sequence: Int64, code_sequence: Int64, abi_sequence: Int64) {
         self.receiver = receiver
         self.act_digest = act_digest
         self.global_sequence = global_sequence
